Gatos - Updates
- Fixed a potential divide by zero issue
- Fixed a typo.  We are now passing in the right object for threshold calc.
- Implemented, then removed, the post-processing routines.  They appeared to greatly reduce detail.  Holes were closed, even when they were supposed to stay open.  Artificially thicker lines as well.

+++ Doxa/Grayscale.hpp
@@ -212,7 +212,7 @@ namespace Doxa
 		/// <summary>
 		/// Linear sRGB to XYZ color space with a D65 white point
 		/// </summary>
-		static constexpr inline XYZ RGBToXYZ(double r, double g, double b)
+		static constexpr inline XYZ RGBToXYZ(float r, float g, float b)
 		{
 			return {
 				0.4124564f * r + 0.3575761f * g + 0.1804375f * b,
